recommendation-agent/requirements.txt
# requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
requests
grpcio
grpcio-tools
protobuf==4.25.3
python-multipart 
